hotel/requirements.txt
asgiref==3.2.3
Babel==2.8.0
Django==3.0.2
django-phonenumber-field==4.0.0
numpy==1.18.1
phonenumbers==8.11.3
pytz==2019.3
sqlparse==0.3.0
virtualenv==16.7.9
